requirements.txt
Django~=4.1.3
stripe
Pillow
python-decouple